se cambio el data type de int a float, los test siguen pasando

+++ src/Product.java
@@ -1,6 +1,6 @@
 
 
 public interface Product {
-	public int total();
+	public double total();
 
 }
